스크립트문과 관리자페이지 오류 수정

+++ Catriend/src/main/java/catriend/command/CatBoardsDeletesCommand.java
@@ -16,14 +16,14 @@ public class CatBoardsDeletesCommand implements CatCommand {
 		// 파라미터 받기
 		Map<String, Object> paramMap = model.asMap();
 		HttpServletRequest req = (HttpServletRequest) paramMap.get("req");
-		
+
 		String[] c_indexs = req.getParameterValues("select");
-		
-		for (String arrStr : c_indexs) {
-			int c_index = Integer.parseInt(arrStr);
-			dao.DeleteCatBoarder(c_index);
+		if (c_indexs != null) {
+			for (String arrStr : c_indexs) {
+				int c_index = Integer.parseInt(arrStr);
+				dao.DeleteCatBoarder(c_index);
+			}
 		}
-		
 		System.out.println("CatBoardsDeletesCommand");
 	}
 }
