+++ Catriend/src/main/java/catriend/command/CatsDeletesCommand.java
@@ -8,7 +8,6 @@ import org.springframework.ui.Model;
 
 import catriend.model.CatsDAO;
 
-
 public class CatsDeletesCommand implements CatCommand {
 	@Override
 	public void execute(Model model) {
@@ -17,14 +16,14 @@ public class CatsDeletesCommand implements CatCommand {
 		// 파라미터 받기
 		Map<String, Object> paramMap = model.asMap();
 		HttpServletRequest req = (HttpServletRequest) paramMap.get("req");
-		
+
 		String[] c_ids = req.getParameterValues("select");
-		
-		for (String arrStr : c_ids) {
-			int c_index = Integer.parseInt(arrStr);
-			dao.DeleteCat(c_index);
+		if (c_ids != null) {
+			for (String arrStr : c_ids) {
+				int c_index = Integer.parseInt(arrStr);
+				dao.DeleteCat(c_index);
+			}
 		}
-		
 		System.out.println("UsersDeletesCommand");
 	}
 }
